Move test next to services
