+++ data-science-lab-app/electron/spec/support/jasmine.json
@@ -1,7 +1,7 @@
 {
     "spec_dir": "./test/spec",
     "spec_files": [
-      "../**/*[sS]pec.js"
+      "../electron/**/*[sS]pec.js"
     ],
     "helpers": [
       "helpers/**/*.js"
