+++ data-science-lab-app/package.json
@@ -7,8 +7,10 @@
     "e2e": "ng e2e",
     "ang:gc": "ng g c ",
     "test:ang": "ng test",
-    "test:elc": "tsc --p tsconfig.electron.spec.json && jasmine --config=./electron/spec/support/jasmine.json || exit 0",
+    "test:elc": "npm run build:elc:test && npm run jasmine:elc && npm run clean:test",
+    "jasmine:elc": "jasmine --config=./electron/spec/support/jasmine.json || exit 0",
     "build:ang": "ng build --base-href ./",
+    "build:elc:test": "tsc --p tsconfig.electron.spec.json",
     "build:elc": "tsc --p tsconfig.electron.json",
     "build:all": "npm run build:elc && npm run build:ang",
     "build:start": "npm run build:all && npm run start",
